@@ -2,7 +2,7 @@ name: audioplayers_test_server
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   shelf: ^1.2.0
