@@ -30,6 +30,7 @@ dependencies:
   flutter:
     sdk: flutter
   http: ^0.13.5
+  meta: ^1.8.0
   path_provider: ^2.0.12
   synchronized: ^3.0.1
   uuid: ^3.0.7
@@ -41,6 +42,6 @@ dev_dependencies:
     sdk: flutter
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
 
