@@ -7,6 +7,7 @@ repository: https://github.com/bluefireteam/audioplayers/tree/master/packages/au
 dependencies:
   flutter:
     sdk: flutter
+  meta: ^1.8.0
   plugin_platform_interface: ^2.0.2
 
 dev_dependencies:
@@ -16,5 +17,5 @@ dev_dependencies:
     sdk: flutter
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
