@@ -1,5 +1,4 @@
 import 'dart:async';
-
 // TODO(gustl22): remove when upgrading min Flutter version to >=3.3.0
 // ignore: unnecessary_import
 import 'dart:typed_data';
@@ -7,6 +6,7 @@ import 'dart:typed_data';
 import 'package:audioplayers/audioplayers.dart';
 import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
 import 'package:flutter/services.dart';
+import 'package:meta/meta.dart';
 import 'package:uuid/uuid.dart';
 
 const _uuid = Uuid();
@@ -19,7 +19,7 @@ const _uuid = Uuid();
 /// hooks for handlers and callbacks.
 class AudioPlayer {
   static final global = GlobalAudioScope();
-  static final _platform = AudioplayersPlatformInterface.instance;
+  final _platform = AudioplayersPlatformInterface.instance;
 
   /// This is the [AudioCache] instance used by this player.
   /// Unless you want to control multiple caches separately, you don't need to
@@ -41,7 +41,21 @@ class AudioPlayer {
     _playerState = state;
   }
 
-  late StreamSubscription _onPlayerCompleteStreamSubscription;
+  /// Completer to wait until the native player and its event stream are
+  /// created.
+  @visibleForTesting
+  final creatingCompleter = Completer<void>();
+
+  late final StreamSubscription _onPlayerCompleteStreamSubscription;
+
+  late final StreamSubscription _onLogStreamSubscription;
+
+  /// Stream controller to be able to get a stream on initialization, before the
+  /// native event stream is ready via [_create] method.
+  final _eventStreamController = StreamController<PlayerEvent>.broadcast();
+  late final StreamSubscription _eventStreamSubscription;
+
+  Stream<PlayerEvent> get eventStream => _eventStreamController.stream;
 
   final StreamController<PlayerState> _playerStateController =
       StreamController<PlayerState>.broadcast();
@@ -55,15 +69,17 @@ class AudioPlayer {
   /// position of the playback if the status is [PlayerState.playing].
   ///
   /// You can use it on a progress bar, for instance.
-  Stream<Duration> get onPositionChanged =>
-      _platform.positionStream.filter(playerId);
+  Stream<Duration> get onPositionChanged => eventStream
+      .where((event) => event.eventType == PlayerEventType.position)
+      .map((event) => event.position!);
 
   /// Stream of changes on audio duration.
   ///
   /// An event is going to be sent as soon as the audio duration is available
   /// (it might take a while to download or buffer it).
-  Stream<Duration> get onDurationChanged =>
-      _platform.durationStream.filter(playerId);
+  Stream<Duration> get onDurationChanged => eventStream
+      .where((event) => event.eventType == PlayerEventType.duration)
+      .map((event) => event.duration!);
 
   /// Stream of player completions.
   ///
@@ -72,13 +88,18 @@ class AudioPlayer {
   ///
   /// [ReleaseMode.loop] also sends events to this stream.
   Stream<void> get onPlayerComplete =>
-      _platform.completeStream.filter(playerId);
+      eventStream.where((event) => event.eventType == PlayerEventType.complete);
 
   /// Stream of seek completions.
   ///
   /// An event is going to be sent as soon as the audio seek is finished.
-  Stream<void> get onSeekComplete =>
-      _platform.seekCompleteStream.filter(playerId);
+  Stream<void> get onSeekComplete => eventStream
+      .where((event) => event.eventType == PlayerEventType.seekComplete);
+
+  /// Stream of log events.
+  Stream<String> get onLog => eventStream
+      .where((event) => event.eventType == PlayerEventType.log)
+      .map((event) => event.logMessage!);
 
   /// An unique ID generated for this instance of [AudioPlayer].
   ///
@@ -97,12 +118,38 @@ class AudioPlayer {
 
   /// Creates a new instance and assigns an unique id to it.
   AudioPlayer({String? playerId}) : playerId = playerId ?? _uuid.v4() {
-    _onPlayerCompleteStreamSubscription = onPlayerComplete.listen((_) {
-      state = PlayerState.completed;
-      if (releaseMode == ReleaseMode.release) {
-        _source = null;
-      }
-    });
+    _onLogStreamSubscription = onLog.listen(
+      (log) => Logger.log('$log\nSource: $_source'),
+      onError: (Object e, [StackTrace? stackTrace]) => Logger.error(
+        AudioPlayerException(this, cause: e),
+        stackTrace,
+      ),
+    );
+    _onPlayerCompleteStreamSubscription = onPlayerComplete.listen(
+      (_) {
+        state = PlayerState.completed;
+        if (releaseMode == ReleaseMode.release) {
+          _source = null;
+        }
+      },
+      onError: (Object _, [StackTrace? __]) {
+        /* Errors are already handled via log stream */
+      },
+    );
+    _create();
+  }
+
+  Future<void> _create() async {
+    try {
+      await _platform.create(playerId);
+      _eventStreamSubscription = _platform.getEventStream(playerId).listen(
+            _eventStreamController.add,
+            onError: _eventStreamController.addError,
+          );
+      creatingCompleter.complete();
+    } on Exception catch (e, st) {
+      creatingCompleter.completeError(e, st);
+    }
   }
 
   Future<void> play(
@@ -132,12 +179,14 @@ class AudioPlayer {
     return resume();
   }
 
-  Future<void> setAudioContext(AudioContext ctx) {
+  Future<void> setAudioContext(AudioContext ctx) async {
+    await creatingCompleter.future;
     return _platform.setAudioContext(playerId, ctx);
   }
 
-  Future<void> setPlayerMode(PlayerMode mode) {
+  Future<void> setPlayerMode(PlayerMode mode) async {
     _mode = mode;
+    await creatingCompleter.future;
     return _platform.setPlayerMode(playerId, mode);
   }
 
@@ -146,6 +195,7 @@ class AudioPlayer {
   /// If you call [resume] later, the audio will resume from the point that it
   /// has been paused.
   Future<void> pause() async {
+    await creatingCompleter.future;
     await _platform.pause(playerId);
     state = PlayerState.paused;
   }
@@ -155,12 +205,14 @@ class AudioPlayer {
   /// The position is going to be reset and you will no longer be able to resume
   /// from the last point.
   Future<void> stop() async {
+    await creatingCompleter.future;
     await _platform.stop(playerId);
     state = PlayerState.stopped;
   }
 
   /// Resumes the audio that has been paused or stopped.
   Future<void> resume() async {
+    await creatingCompleter.future;
     await _platform.resume(playerId);
     state = PlayerState.playing;
   }
@@ -170,13 +222,15 @@ class AudioPlayer {
   /// The resources are going to be fetched or buffered again as soon as you
   /// call [resume] or change the source.
   Future<void> release() async {
+    await creatingCompleter.future;
     await _platform.release(playerId);
     state = PlayerState.stopped;
     _source = null;
   }
 
   /// Moves the cursor to the desired position.
-  Future<void> seek(Duration position) {
+  Future<void> seek(Duration position) async {
+    await creatingCompleter.future;
     return _platform.seek(playerId, position);
   }
 
@@ -185,7 +239,8 @@ class AudioPlayer {
   /// -1 - The left channel is at full volume; the right channel is silent.
   ///  1 - The right channel is at full volume; the left channel is silent.
   ///  0 - Both channels are at the same volume.
-  Future<void> setBalance(double balance) {
+  Future<void> setBalance(double balance) async {
+    await creatingCompleter.future;
     return _platform.setBalance(playerId, balance);
   }
 
@@ -193,15 +248,17 @@ class AudioPlayer {
   ///
   /// 0 is mute and 1 is the max volume. The values between 0 and 1 are linearly
   /// interpolated.
-  Future<void> setVolume(double volume) {
+  Future<void> setVolume(double volume) async {
+    await creatingCompleter.future;
     return _platform.setVolume(playerId, volume);
   }
 
   /// Sets the release mode.
   ///
   /// Check [ReleaseMode]'s doc to understand the difference between the modes.
-  Future<void> setReleaseMode(ReleaseMode releaseMode) {
+  Future<void> setReleaseMode(ReleaseMode releaseMode) async {
     _releaseMode = releaseMode;
+    await creatingCompleter.future;
     return _platform.setReleaseMode(playerId, releaseMode);
   }
 
@@ -209,7 +266,8 @@ class AudioPlayer {
   ///
   /// iOS and macOS have limits between 0.5 and 2x
   /// Android SDK version should be 23 or higher
-  Future<void> setPlaybackRate(double playbackRate) {
+  Future<void> setPlaybackRate(double playbackRate) async {
+    await creatingCompleter.future;
     return _platform.setPlaybackRate(playerId, playbackRate);
   }
 
@@ -217,7 +275,8 @@ class AudioPlayer {
   ///
   /// This will delegate to one of the specific methods below depending on
   /// the source type.
-  Future<void> setSource(Source source) {
+  Future<void> setSource(Source source) async {
+    await creatingCompleter.future;
     return source.setOnPlayer(this);
   }
 
@@ -225,8 +284,9 @@ class AudioPlayer {
   ///
   /// The resources will start being fetched or buffered as soon as you call
   /// this method.
-  Future<void> setSourceUrl(String url) {
+  Future<void> setSourceUrl(String url) async {
     _source = UrlSource(url);
+    await creatingCompleter.future;
     return _platform.setSourceUrl(playerId, url, isLocal: false);
   }
 
@@ -234,8 +294,9 @@ class AudioPlayer {
   ///
   /// The resources will start being fetched or buffered as soon as you call
   /// this method.
-  Future<void> setSourceDeviceFile(String path) {
+  Future<void> setSourceDeviceFile(String path) async {
     _source = DeviceFileSource(path);
+    await creatingCompleter.future;
     return _platform.setSourceUrl(playerId, path, isLocal: true);
   }
 
@@ -247,11 +308,13 @@ class AudioPlayer {
   Future<void> setSourceAsset(String path) async {
     _source = AssetSource(path);
     final url = await audioCache.load(path);
+    await creatingCompleter.future;
     return _platform.setSourceUrl(playerId, url.path, isLocal: true);
   }
 
-  Future<void> setSourceBytes(Uint8List bytes) {
+  Future<void> setSourceBytes(Uint8List bytes) async {
     _source = BytesSource(bytes);
+    await creatingCompleter.future;
     return _platform.setSourceBytes(playerId, bytes);
   }
 
@@ -261,6 +324,7 @@ class AudioPlayer {
   /// It will be available as soon as the audio duration is available
   /// (it might take a while to download or buffer it if file is not local).
   Future<Duration?> getDuration() async {
+    await creatingCompleter.future;
     final milliseconds = await _platform.getDuration(playerId);
     if (milliseconds == null) {
       return null;
@@ -270,6 +334,7 @@ class AudioPlayer {
 
   // Gets audio current playing position
   Future<Duration?> getCurrentPosition() async {
+    await creatingCompleter.future;
     final milliseconds = await _platform.getCurrentPosition(playerId);
     if (milliseconds == null) {
       return null;
@@ -286,12 +351,17 @@ class AudioPlayer {
     await release();
 
     final futures = <Future>[
+      creatingCompleter.future,
       if (!_playerStateController.isClosed) _playerStateController.close(),
-      _onPlayerCompleteStreamSubscription.cancel()
+      _onPlayerCompleteStreamSubscription.cancel(),
+      _onLogStreamSubscription.cancel(),
+      _eventStreamSubscription.cancel(),
+      _eventStreamController.close(),
     ];
 
     _source = null;
 
     await Future.wait<dynamic>(futures);
+    await _platform.dispose(playerId);
   }
 }
