@@ -197,7 +197,9 @@ class _ExampleAppState extends State<ExampleApp> {
                               TabData(
                                 key: 'loggerTab',
                                 label: 'Log',
-                                content: const LoggerTab(),
+                                content: LoggerTab(
+                                  player: player,
+                                ),
                               ),
                             ],
                           ),
