@@ -1,7 +1,7 @@
 export 'package:audioplayers_platform_interface/src/api/audio_context.dart';
 export 'package:audioplayers_platform_interface/src/api/audio_context_config.dart';
-export 'package:audioplayers_platform_interface/src/api/for_player.dart';
-export 'package:audioplayers_platform_interface/src/api/log_level.dart';
+export 'package:audioplayers_platform_interface/src/api/global_event.dart';
+export 'package:audioplayers_platform_interface/src/api/player_event.dart';
 export 'package:audioplayers_platform_interface/src/api/player_mode.dart';
 export 'package:audioplayers_platform_interface/src/api/player_state.dart';
 export 'package:audioplayers_platform_interface/src/api/release_mode.dart';
@@ -10,4 +10,6 @@ export 'src/audio_cache.dart';
 export 'src/audio_pool.dart';
 export 'src/audioplayer.dart';
 export 'src/global_audio_scope.dart';
+export 'src/log_level.dart';
+export 'src/logger.dart';
 export 'src/source.dart';
