@@ -15,96 +15,140 @@ let GLOBAL_CHANNEL_NAME = "xyz.luan/audioplayers.global"
 
 public class SwiftAudioplayersDarwinPlugin: NSObject, FlutterPlugin {
     var registrar: FlutterPluginRegistrar
-    var channel: FlutterMethodChannel
-    var globalChannel: FlutterMethodChannel
-    
+    var binaryMessenger: FlutterBinaryMessenger
+    var methods: FlutterMethodChannel
+    var globalMethods: FlutterMethodChannel
+    var globalEvents: GlobalAudioPlayersStreamHandler
+
     var globalContext = AudioContext()
-    var players = [String : WrappedMediaPlayer]()
-    
-    init(registrar: FlutterPluginRegistrar, channel: FlutterMethodChannel, globalChannel: FlutterMethodChannel) {
+    var players = [String: WrappedMediaPlayer]()
+
+    init(registrar: FlutterPluginRegistrar,
+         binaryMessenger: FlutterBinaryMessenger,
+         methodChannel: FlutterMethodChannel,
+         globalMethodChannel: FlutterMethodChannel,
+         globalEventChannel: FlutterEventChannel) {
         self.registrar = registrar
-        self.channel = channel
-        self.globalChannel = globalChannel
-        
-        globalContext.apply()
-        
+        self.binaryMessenger = binaryMessenger
+        self.methods = methodChannel
+        self.globalMethods = globalMethodChannel
+        self.globalEvents = GlobalAudioPlayersStreamHandler()
+
+        do {
+            try globalContext.apply()
+        } catch {
+            // ignore error on initialization
+        }
+
         super.init()
+
+        self.globalMethods.setMethodCallHandler(self.handleGlobalMethodCall)
+        globalEventChannel.setStreamHandler(self.globalEvents);
     }
-    
+
     public static func register(with registrar: FlutterPluginRegistrar) {
         // apparently there is a bug in Flutter causing some inconsistency between Flutter and FlutterMacOS
+        // See: https://github.com/flutter/flutter/issues/118103
         #if os(iOS)
         let binaryMessenger = registrar.messenger()
         #else
         let binaryMessenger = registrar.messenger
         #endif
-        
-        let channel = FlutterMethodChannel(name: CHANNEL_NAME, binaryMessenger: binaryMessenger)
-        let globalChannel = FlutterMethodChannel(name: GLOBAL_CHANNEL_NAME, binaryMessenger: binaryMessenger)
 
-        let instance = SwiftAudioplayersDarwinPlugin(registrar: registrar, channel: channel, globalChannel: globalChannel)
-        registrar.addMethodCallDelegate(instance, channel: channel)
-        registrar.addMethodCallDelegate(instance, channel: globalChannel)
+        let methods = FlutterMethodChannel(name: CHANNEL_NAME, binaryMessenger: binaryMessenger)
+        let globalMethods = FlutterMethodChannel(name: GLOBAL_CHANNEL_NAME, binaryMessenger: binaryMessenger)
+        let globalEvents = FlutterEventChannel(name: GLOBAL_CHANNEL_NAME + "/events", binaryMessenger: binaryMessenger)
+
+        let instance = SwiftAudioplayersDarwinPlugin(
+                registrar: registrar,
+                binaryMessenger: binaryMessenger,
+                methodChannel: methods,
+                globalMethodChannel: globalMethods,
+                globalEventChannel: globalEvents)
+        registrar.addMethodCallDelegate(instance, channel: methods)
     }
 
     public func detachFromEngine(for registrar: FlutterPluginRegistrar) {
         dispose()
     }
-    
+
     func dispose() {
         for (_, player) in self.players {
             player.dispose()
         }
         self.players = [:]
     }
-    
-    public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+
+    private func handleGlobalMethodCall(call: FlutterMethodCall, result: @escaping FlutterResult) {
         let method = call.method
-        
+
         guard let args = call.arguments as? [String: Any] else {
-            Logger.error("Failed to parse call.arguments from Flutter.")
-            result(0)
+            result(FlutterError(code: "DarwinAudioError", message: "Failed to parse call.arguments from Flutter.", details: nil))
             return
         }
 
-        Logger.info("method: %@", method)
-
         // global handlers (no playerId)
-        if method == "changeLogLevel" {
-            guard let valueName = args["value"] as! String? else {
-                Logger.error("Null value received on changeLogLevel")
-                result(0)
-                return
-            }
-            guard let value = LogLevel.parse(valueName) else {
-                Logger.error("Invalid value received on changeLogLevel")
-                result(0)
-                return
-            }
+        if method == "setAudioContext" {
+            do {
+                guard let context = try AudioContext.parse(args: args) else {
+                    result(FlutterError(code: "DarwinAudioError", message: "Error calling setAudioContext, context could not be parsed", details: nil))
+                    return
+                }
+                globalContext = context
 
-            Logger.logLevel = value
-            result(1)
-            return
-        } else if method == "setGlobalAudioContext" {
-            guard let context = AudioContext.parse(args: args) else {
-                result(0)
+                try globalContext.apply()
+            } catch AudioPlayerError.warning(let warnMsg) {
+                globalEvents.onLog(message: warnMsg)
+            }  catch {
+                result(FlutterError(code: "DarwinAudioError", message: "Error configuring global audio session: \(error)", details: nil))
+            }
+        } else if method == "emitLog" {
+            guard let message = args["message"] as? String else {
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling emitLog, message cannot be null", details: nil))
                 return
             }
-            globalContext = context
-            globalContext.apply()
-            result(1)
+            globalEvents.onLog(message: message)
+        } else if method == "emitError" {
+            guard let code = args["code"] as? String else {
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling emitError, code cannot be null", details: nil))
+                return
+            }
+            guard let message = args["message"] as? String else {
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling emitError, message cannot be null", details: nil))
+                return
+            }
+            globalEvents.onError(code: code, message: message, details: nil)
+        } else {
+            result(FlutterMethodNotImplemented)
+            return
+        }
+
+        // default result (bypass by adding `return` to your branch)
+        result(1)
+    }
+
+    public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+        let method = call.method
+
+        guard let args = call.arguments as? [String: Any] else {
+            result(FlutterError(code: "DarwinAudioError", message: "Failed to parse call.arguments from Flutter.", details: nil))
             return
         }
 
         // player specific handlers
         guard let playerId = args["playerId"] as? String else {
-            Logger.error("Call missing mandatory parameter playerId.")
-            result(0)
+            result(FlutterError(code: "DarwinAudioError", message: "Call missing mandatory parameter playerId.", details: nil))
             return
         }
-        Logger.info("playerId: %@", playerId)
-        let player = self.getOrCreatePlayer(playerId: playerId)
-        
+
+        if method == "create" {
+            self.createPlayer(playerId: playerId)
+            result(1)
+            return
+        }
+
+        let player = self.getPlayer(playerId: playerId)
+
         if method == "pause" {
             player.pause()
         } else if method == "resume" {
@@ -118,8 +162,7 @@ public class SwiftAudioplayersDarwinPlugin: NSObject, FlutterPlugin {
             player.release()
         } else if method == "seek" {
             guard let position = args["position"] as? Int else {
-                Logger.error("Null position received on seek")
-                result(0)
+                result(FlutterError(code: "DarwinAudioError", message: "Null position received on seek", details: nil))
                 return
             }
             let time = toCMTime(millis: position)
@@ -130,51 +173,48 @@ public class SwiftAudioplayersDarwinPlugin: NSObject, FlutterPlugin {
         } else if method == "setSourceUrl" {
             let url: String? = args["url"] as? String
             let isLocal: Bool = (args["isLocal"] as? Bool) ?? false
-            
+
             if url == nil {
-                Logger.error("Null URL received on setSourceUrl")
-                result(0)
+                result(FlutterError(code: "DarwinAudioError", message: "Null URL received on setSourceUrl", details: nil))
                 return
             }
-            
-            player.setSourceUrl(url: url!, isLocal: isLocal) {
+
+            player.setSourceUrl(url: url!, isLocal: isLocal, completer: {
                 result(1)
-            }
+            }, completerError: {
+                result(FlutterError(code: "DarwinAudioError", message: "AVPlayerItem.Status.failed on setSourceUrl", details: nil))
+            })
             return
         } else if method == "setSourceBytes" {
-            Logger.error("setSourceBytes is not currently implemented on iOS")
-            result(0)
+            result(FlutterError(code: "DarwinAudioError", message: "setSourceBytes is not currently implemented on iOS", details: nil))
             return
         } else if method == "getDuration" {
             let duration = player.getDuration()
             result(duration)
         } else if method == "setVolume" {
             guard let volume = args["volume"] as? Double else {
-                Logger.error("Error calling setVolume, volume cannot be null")
-                result(0)
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling setVolume, volume cannot be null", details: nil))
                 return
             }
-            
+
             player.setVolume(volume: volume)
         } else if method == "setBalance" {
-            Logger.error("setBalance is not currently implemented on iOS")
+            player.eventHandler.onLog(message: "setBalance is not currently implemented on iOS")
             result(0)
             return
-       } else if method == "getCurrentPosition" {
+        } else if method == "getCurrentPosition" {
             let currentPosition = player.getCurrentPosition()
             result(currentPosition)
             return
         } else if method == "setPlaybackRate" {
             guard let playbackRate = args["playbackRate"] as? Double else {
-                Logger.error("Error calling setPlaybackRate, playbackRate cannot be null")
-                result(0)
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling setPlaybackRate, playbackRate cannot be null", details: nil))
                 return
             }
             player.setPlaybackRate(playbackRate: playbackRate)
         } else if method == "setReleaseMode" {
             guard let releaseMode = args["releaseMode"] as? String else {
-                Logger.error("Error calling setReleaseMode, releaseMode cannot be null")
-                result(0)
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling setReleaseMode, releaseMode cannot be null", details: nil))
                 return
             }
             // Note: there is no "release" on iOS; hence we only care if it's looping or not
@@ -183,57 +223,147 @@ public class SwiftAudioplayersDarwinPlugin: NSObject, FlutterPlugin {
         } else if method == "setPlayerMode" {
             // no-op for darwin; only one player mode
         } else if method == "setAudioContext" {
-            Logger.info("iOS does not allow for player-specific audio contexts; `setAudioContext` will set the global audio context instead (like `setGlobalAudioContext`).")
-            guard let context = AudioContext.parse(args: args) else {
-                result(0)
+            player.eventHandler.onLog(message: "iOS does not allow for player-specific audio contexts; `setAudioContext` will set the global audio context instead (like `global.setAudioContext`).")
+            do {
+                guard let context = try AudioContext.parse(args: args) else {
+                    result(FlutterError(code: "DarwinAudioError", message: "Error calling setAudioContext, context could not be parsed", details: nil))
+                    return
+                }
+                globalContext = context
+
+                try globalContext.apply()
+            } catch AudioPlayerError.warning(let warnMsg) {
+                globalEvents.onLog(message: warnMsg)
+            } catch {
+                result(FlutterError(code: "DarwinAudioError", message: "Error configuring audio session: \(error)", details: nil))
+            }
+        } else if method == "emitLog" {
+            guard let message = args["message"] as? String else {
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling emitLog, message cannot be null", details: nil))
                 return
             }
-            globalContext = context
-            globalContext.apply()
+            player.eventHandler.onLog(message: message)
+        } else if method == "emitError" {
+            guard let code = args["code"] as? String else {
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling emitError, code cannot be null", details: nil))
+                return
+            }
+            guard let message = args["message"] as? String else {
+                result(FlutterError(code: "DarwinAudioError", message: "Error calling emitError, message cannot be null", details: nil))
+                return
+            }
+            player.eventHandler.onError(code: code, message: message, details: nil)
         } else {
-            Logger.error("Called not implemented method: %@", method)
             result(FlutterMethodNotImplemented)
             return
         }
-        
+
         // default result (bypass by adding `return` to your branch)
         result(1)
     }
-    
-    func getOrCreatePlayer(playerId: String) -> WrappedMediaPlayer {
-        if let player = players[playerId] {
-            return player
-        }
+
+    func createPlayer(playerId: String) {
+        let eventChannel = FlutterEventChannel(name: CHANNEL_NAME + "/events/" + playerId, binaryMessenger: self.binaryMessenger)
+
+        let eventHandler = AudioPlayersStreamHandler()
+
+        eventChannel.setStreamHandler(eventHandler);
+
         let newPlayer = WrappedMediaPlayer(
-            reference: self,
-            playerId: playerId
+                reference: self,
+                eventHandler: eventHandler
         )
         players[playerId] = newPlayer
-        return newPlayer
     }
-    
-    func onSeekComplete(playerId: String, finished: Bool) {
-        channel.invokeMethod("audio.onSeekComplete", arguments: ["playerId": playerId, "value": finished])
+
+    func getPlayer(playerId: String) -> WrappedMediaPlayer {
+        return players[playerId]!
     }
-    
-    func onComplete(playerId: String) {
-        channel.invokeMethod("audio.onComplete", arguments: ["playerId": playerId])
-    }
-    
-    func onCurrentPosition(playerId: String, millis: Int) {
-        channel.invokeMethod("audio.onCurrentPosition", arguments: ["playerId": playerId, "value": millis])
-    }
-    
-    func onError(playerId: String) {
-        channel.invokeMethod("audio.onError", arguments: ["playerId": playerId, "value": "AVPlayerItem.Status.failed"])
-    }
-    
-    func onDuration(playerId: String, millis: Int) {
-        channel.invokeMethod("audio.onDuration", arguments: ["playerId": playerId, "value": millis])
-    }
-    
+
     func controlAudioSession() {
-        let anyIsPlaying = players.values.contains { player in player.isPlaying }
-        globalContext.activateAudioSession(active: anyIsPlaying)
+        let anyIsPlaying = players.values.contains { player in
+            player.isPlaying
+        }
+
+        do {
+            try globalContext.activateAudioSession(active: anyIsPlaying)
+        } catch {
+            self.globalEvents.onError(code: "DarwinAudioError", message: "Error configuring audio session: \(error)", details: nil)
+        }
+    }
+}
+
+class AudioPlayersStreamHandler: NSObject, FlutterStreamHandler {
+    var sink: FlutterEventSink?
+
+    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
+        self.sink = events
+        // events(FlutterEndOfEventStream) // when stream is over
+        return nil
+    }
+
+    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
+        return nil
+    }
+
+    func onSeekComplete() {
+        if let eventSink = self.sink {
+            eventSink(["event": "audio.onSeekComplete"])
+        }
+    }
+
+    func onComplete() {
+        if let eventSink = self.sink {
+            eventSink(["event": "audio.onComplete"])
+        }
+    }
+
+    func onCurrentPosition(millis: Int) {
+        if let eventSink = self.sink {
+            eventSink(["event": "audio.onCurrentPosition", "value": millis])
+        }
+    }
+
+    func onDuration(millis: Int) {
+        if let eventSink = self.sink {
+            eventSink(["event": "audio.onDuration", "value": millis])
+        }
+    }
+
+    func onLog(message: String) {
+        if let eventSink = self.sink {
+            eventSink(["event": "audio.onLog", "value": message])
+        }
+    }
+
+    func onError(code: String, message: String, details: Any?) {
+        if let eventSink = self.sink {
+            eventSink(FlutterError(code: code, message: message, details: details))
+        }
+    }
+}
+
+class GlobalAudioPlayersStreamHandler: NSObject, FlutterStreamHandler {
+    var sink: FlutterEventSink?
+
+    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
+        self.sink = events
+        return nil
+    }
+
+    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
+        return nil
+    }
+
+    func onLog(message: String) {
+        if let eventSink = self.sink {
+            eventSink(["event": "audio.onLog", "value": message])
+        }
+    }
+
+    func onError(code: String, message: String, details: Any?) {
+        if let eventSink = self.sink {
+            eventSink(FlutterError(code: code, message: message, details: details))
+        }
     }
 }
