@@ -2,13 +2,13 @@ import MediaPlayer
 
 // no-op impl of AudioContext for macos
 struct AudioContext {
-    func activateAudioSession(active: Bool) {}
+    func activateAudioSession(active: Bool) throws {}
 
-    func apply() {
-        Logger.error("AudioContext configuration is not available on macOS")
+    func apply() throws {
+        throw AudioPlayerError.warning("AudioContext configuration is not available on macOS")
     }
 
-    static func parse(args: [String: Any]) -> AudioContext? {
+    static func parse(args: [String: Any]) throws -> AudioContext? {
         return AudioContext()
     }
 }
