@@ -20,8 +20,9 @@ extern "C" {
 }
 
 class AudioPlayer {
-public:
-    AudioPlayer(std::string playerId, FlMethodChannel *channel);
+   public:
+    AudioPlayer(std::string playerId, FlMethodChannel *methodChannel,
+                FlEventChannel *eventChannel);
 
     int64_t GetPosition();
 
@@ -49,9 +50,14 @@ public:
 
     void SetSourceUrl(std::string url);
 
+    void OnError(const gchar *code, const gchar *message, FlValue *details,
+                 GError **error);
+
+    void OnLog(const gchar *message);
+
     virtual ~AudioPlayer();
 
-private:
+   private:
     // Gst members
     GstElement *playbin;
     GstElement *source;
@@ -66,7 +72,8 @@ private:
 
     std::string _url{};
     std::string _playerId;
-    FlMethodChannel *_channel;
+    FlMethodChannel *_methodChannel;
+    FlEventChannel *_eventChannel;
 
     static void SourceSetup(GstElement *playbin, GstElement *source,
                             GstElement **p_src);
