@@ -1,32 +1,71 @@
 import 'package:audioplayers_platform_interface/src/api/audio_context.dart';
-import 'package:audioplayers_platform_interface/src/api/log_level.dart';
+import 'package:audioplayers_platform_interface/src/api/global_event.dart';
 import 'package:audioplayers_platform_interface/src/global_audioplayers_platform_interface.dart';
-import 'package:audioplayers_platform_interface/src/method_channel_interface.dart';
+import 'package:audioplayers_platform_interface/src/map_extension.dart';
+import 'package:audioplayers_platform_interface/src/method_channel_extension.dart';
 import 'package:flutter/services.dart';
 
-class GlobalAudioplayersPlatform extends GlobalAudioplayersPlatformInterface {
-  static const MethodChannel _channel =
+class GlobalAudioplayersPlatform extends GlobalAudioplayersPlatformInterface
+    with
+        MethodChannelGlobalAudioplayersPlatform,
+        EventChannelGlobalAudioplayersPlatform {
+  GlobalAudioplayersPlatform();
+}
+
+mixin MethodChannelGlobalAudioplayersPlatform
+    implements MethodChannelGlobalAudioplayersPlatformInterface {
+  static const MethodChannel _globalMethodChannel =
       MethodChannel('xyz.luan/audioplayers.global');
-
-  static LogLevel _logLevel = LogLevel.error;
-
-  @override
-  Future<void> changeLogLevel(LogLevel value) {
-    _logLevel = value;
-    return _channel.call(
-      'changeLogLevel',
-      <String, dynamic>{'value': value.toString()},
-    );
-  }
 
   @override
   Future<void> setGlobalAudioContext(AudioContext ctx) {
-    return _channel.call(
-      'setGlobalAudioContext',
+    return _globalMethodChannel.call(
+      'setAudioContext',
       ctx.toJson(),
     );
   }
 
   @override
-  LogLevel get logLevel => _logLevel;
+  Future<void> emitGlobalLog(String message) {
+    return _globalMethodChannel.call(
+      'emitLog',
+      <String, dynamic>{
+        'message': message,
+      },
+    );
+  }
+
+  @override
+  Future<void> emitGlobalError(String code, String message) {
+    return _globalMethodChannel.call(
+      'emitError',
+      <String, dynamic>{
+        'code': code,
+        'message': message,
+      },
+    );
+  }
+}
+
+mixin EventChannelGlobalAudioplayersPlatform
+    implements EventChannelGlobalAudioplayersPlatformInterface {
+  static const _globalEventChannel =
+      EventChannel('xyz.luan/audioplayers.global/events');
+
+  @override
+  Stream<GlobalEvent> getGlobalEventStream() {
+    return _globalEventChannel.receiveBroadcastStream().map((dynamic event) {
+      final map = event as Map<dynamic, dynamic>;
+      final eventType = map.getString('event');
+      switch (eventType) {
+        case 'audio.onLog':
+          final value = map.getString('value');
+          return GlobalEvent(eventType: GlobalEventType.log, logMessage: value);
+        default:
+          throw UnimplementedError(
+            'Global Event Method does not exist $eventType',
+          );
+      }
+    });
+  }
 }
