@@ -1,13 +1,13 @@
 import 'dart:async';
-import 'dart:typed_data';
 
 import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
 import 'package:audioplayers_web/global_audioplayers_web.dart';
 import 'package:audioplayers_web/num_extension.dart';
 import 'package:audioplayers_web/wrapped_player.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
-class AudioplayersPlugin with StreamsInterface {
+class AudioplayersPlugin {
   /// The entrypoint called by the generated plugin registrant.
   static void registerWith(Registrar registrar) {
     AudioplayersPlatformInterface.instance = WebAudioplayersPlatform();
@@ -16,18 +16,27 @@ class AudioplayersPlugin with StreamsInterface {
   }
 }
 
-class WebAudioplayersPlatform extends AudioplayersPlatformInterface
-    with StreamsInterface {
+class WebAudioplayersPlatform extends AudioplayersPlatformInterface {
   // players by playerId
   Map<String, WrappedPlayer> players = {};
 
-  WrappedPlayer getOrCreatePlayer(String playerId) {
-    return players.putIfAbsent(playerId, () => WrappedPlayer(playerId, this));
+  @override
+  Future<void> create(String playerId) async {
+    players[playerId] = WrappedPlayer(playerId);
+  }
+
+  WrappedPlayer getPlayer(String playerId) {
+    return players[playerId] != null
+        ? players[playerId]!
+        : throw PlatformException(
+            code: 'WebAudioError',
+            message: 'Player with id $playerId was not created!',
+          );
   }
 
   @override
   Future<int?> getCurrentPosition(String playerId) async {
-    final position = getOrCreatePlayer(playerId).player?.currentTime;
+    final position = getPlayer(playerId).player?.currentTime;
     if (position == null) {
       return null;
     }
@@ -36,7 +45,7 @@ class WebAudioplayersPlatform extends AudioplayersPlatformInterface
 
   @override
   Future<int?> getDuration(String playerId) async {
-    final jsDuration = getOrCreatePlayer(playerId).player?.duration;
+    final jsDuration = getPlayer(playerId).player?.duration;
     if (jsDuration == null) {
       return null;
     }
@@ -45,22 +54,22 @@ class WebAudioplayersPlatform extends AudioplayersPlatformInterface
 
   @override
   Future<void> pause(String playerId) async {
-    getOrCreatePlayer(playerId).pause();
+    getPlayer(playerId).pause();
   }
 
   @override
   Future<void> release(String playerId) async {
-    getOrCreatePlayer(playerId).release();
+    getPlayer(playerId).release();
   }
 
   @override
   Future<void> resume(String playerId) async {
-    await getOrCreatePlayer(playerId).resume();
+    await getPlayer(playerId).resume();
   }
 
   @override
   Future<void> seek(String playerId, Duration position) async {
-    getOrCreatePlayer(playerId).seek(position.inMilliseconds);
+    getPlayer(playerId).seek(position.inMilliseconds);
   }
 
   @override
@@ -68,7 +77,12 @@ class WebAudioplayersPlatform extends AudioplayersPlatformInterface
     String playerId,
     AudioContext audioContext,
   ) async {
-    // no-op: web doesn't have any audio context
+    getPlayer(playerId).eventStreamController.add(
+          const PlayerEvent(
+            eventType: PlayerEventType.log,
+            logMessage: 'Setting AudioContext is not supported on Web',
+          ),
+        );
   }
 
   @override
@@ -81,12 +95,12 @@ class WebAudioplayersPlatform extends AudioplayersPlatformInterface
 
   @override
   Future<void> setPlaybackRate(String playerId, double playbackRate) async {
-    getOrCreatePlayer(playerId).setPlaybackRate(playbackRate);
+    getPlayer(playerId).setPlaybackRate(playbackRate);
   }
 
   @override
   Future<void> setReleaseMode(String playerId, ReleaseMode releaseMode) async {
-    getOrCreatePlayer(playerId).setReleaseMode(releaseMode);
+    getPlayer(playerId).setReleaseMode(releaseMode);
   }
 
   @override
@@ -95,7 +109,7 @@ class WebAudioplayersPlatform extends AudioplayersPlatformInterface
     String url, {
     bool? isLocal,
   }) async {
-    await getOrCreatePlayer(playerId).setUrl(url);
+    await getPlayer(playerId).setUrl(url);
   }
 
   @override
@@ -106,16 +120,41 @@ class WebAudioplayersPlatform extends AudioplayersPlatformInterface
 
   @override
   Future<void> setVolume(String playerId, double volume) async {
-    getOrCreatePlayer(playerId).setVolume(volume);
+    getPlayer(playerId).setVolume(volume);
   }
 
   @override
   Future<void> setBalance(String playerId, double balance) async {
-    getOrCreatePlayer(playerId).setBalance(balance);
+    getPlayer(playerId).setBalance(balance);
   }
 
   @override
   Future<void> stop(String playerId) async {
-    getOrCreatePlayer(playerId).stop();
+    getPlayer(playerId).stop();
+  }
+
+  @override
+  Future<void> emitLog(String playerId, String message) async {
+    getPlayer(playerId).log(message);
+  }
+
+  @override
+  Future<void> emitError(String playerId, String code, String message) async {
+    getPlayer(playerId)
+        .eventStreamController
+        .addError(PlatformException(code: code, message: message));
+  }
+
+  @override
+  Stream<PlayerEvent> getEventStream(String playerId) {
+    return getPlayer(playerId).eventStreamController.stream;
+  }
+
+  @override
+  Future<void> dispose(String playerId) async {
+    await Future.forEach<WrappedPlayer>(
+      players.values,
+      (player) => player.dispose(),
+    );
   }
 }
