@@ -1,22 +1,37 @@
 import 'dart:async';
 
 import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
+import 'package:flutter/services.dart';
 
 class WebGlobalAudioplayersPlatform
     extends GlobalAudioplayersPlatformInterface {
-  // Web implementation currently does not log anything
-  LogLevel _level = LogLevel.error;
-
-  @override
-  Future<void> changeLogLevel(LogLevel value) async {
-    _level = value;
-  }
-
-  @override
-  LogLevel get logLevel => _level;
+  final _eventStreamController = StreamController<GlobalEvent>.broadcast();
 
   @override
   Future<void> setGlobalAudioContext(AudioContext ctx) async {
-    // no-op: web does not support changing audio context
+    _eventStreamController.add(
+      const GlobalEvent(
+        eventType: GlobalEventType.log,
+        logMessage: 'Setting global AudioContext is not supported on Web',
+      ),
+    );
+  }
+
+  @override
+  Stream<GlobalEvent> getGlobalEventStream() {
+    return _eventStreamController.stream;
+  }
+
+  @override
+  Future<void> emitGlobalLog(String message) async {
+    _eventStreamController.add(
+      GlobalEvent(eventType: GlobalEventType.log, logMessage: message),
+    );
+  }
+
+  @override
+  Future<void> emitGlobalError(String code, String message) async {
+    _eventStreamController
+        .addError(PlatformException(code: code, message: message));
   }
 }
