@@ -5,89 +5,100 @@
 
 // For getPlatformVersion; remove unless needed for your plugin implementation.
 #include <VersionHelpers.h>
-
+#include <flutter/encodable_value.h>
+#include <flutter/event_channel.h>
 #include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
 #include <flutter/standard_method_codec.h>
-#include <flutter/encodable_value.h>
-
-#include "audio_player.h"
-#include "Logger.h"
 
 #include <map>
 #include <memory>
 #include <sstream>
 
+#include "audio_player.h"
+
 namespace {
 
 using namespace flutter;
 
-template<typename T>
-T GetArgument(const std::string arg, const EncodableValue* args, T fallback) {
-  T result {fallback};
-  const auto* arguments = std::get_if<EncodableMap>(args);
-  if (arguments) {
-    auto result_it = arguments->find(EncodableValue(arg));
-    if (result_it != arguments->end()) {
-      if(!result_it->second.IsNull())
-        result = std::get<T>(result_it->second);
+template <typename T>
+T GetArgument(const std::string arg, const EncodableValue *args, T fallback) {
+    T result{fallback};
+    const auto *arguments = std::get_if<EncodableMap>(args);
+    if (arguments) {
+        auto result_it = arguments->find(EncodableValue(arg));
+        if (result_it != arguments->end()) {
+            if (!result_it->second.IsNull())
+                result = std::get<T>(result_it->second);
+        }
     }
-  }
-  return result;
+    return result;
 }
 
 class AudioplayersWindowsPlugin : public Plugin {
- public:
-  static void RegisterWithRegistrar(PluginRegistrarWindows *registrar);
+   public:
+    static void RegisterWithRegistrar(PluginRegistrarWindows *registrar);
 
-  AudioplayersWindowsPlugin();
+    AudioplayersWindowsPlugin();
 
-  virtual ~AudioplayersWindowsPlugin();
+    virtual ~AudioplayersWindowsPlugin();
 
- private:
+   private:
+    std::map<std::string, std::unique_ptr<AudioPlayer>> audioPlayers;
 
-  std::map<std::string, std::unique_ptr<AudioPlayer>> audioPlayers;
+    static inline BinaryMessenger *binaryMessenger;
+    static inline std::unique_ptr<MethodChannel<EncodableValue>> methods{};
+    static inline std::unique_ptr<MethodChannel<EncodableValue>>
+        globalMethods{};
+    static inline std::unique_ptr<EventStreamHandler<>> globalEvents{};
 
-  static inline std::unique_ptr<MethodChannel<EncodableValue>> channel{};
-  static inline std::unique_ptr<MethodChannel<EncodableValue>> globalChannel{};
+    // Called when a method is called on this plugin's channel from Dart.
+    void HandleMethodCall(const MethodCall<EncodableValue> &method_call,
+                          std::unique_ptr<MethodResult<EncodableValue>> result);
 
-  // Called when a method is called on this plugin's channel from Dart.
-  void HandleMethodCall(
-      const MethodCall<EncodableValue> &method_call,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+    void HandleGlobalMethodCall(
+        const MethodCall<EncodableValue> &method_call,
+        std::unique_ptr<MethodResult<EncodableValue>> result);
 
-  void HandleGlobalMethodCall(
-      const MethodCall<EncodableValue> &method_call,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+    void CreatePlayer(std::string playerId);
 
-  AudioPlayer* GetPlayer(std::string playerId, std::string mode);
+    AudioPlayer *GetPlayer(std::string playerId);
+
+    void OnGlobalLog(const std::string& message);
 };
 
 // static
 void AudioplayersWindowsPlugin::RegisterWithRegistrar(
     PluginRegistrarWindows *registrar) {
-  channel =
-      std::make_unique<MethodChannel<EncodableValue>>(
-          registrar->messenger(), "xyz.luan/audioplayers",
-          &StandardMethodCodec::GetInstance());
-  globalChannel =
-      std::make_unique<MethodChannel<EncodableValue>>(
-          registrar->messenger(), "xyz.luan/audioplayers.global",
-          &StandardMethodCodec::GetInstance());
+    binaryMessenger = registrar->messenger();
+    methods = std::make_unique<MethodChannel<EncodableValue>>(
+        binaryMessenger, "xyz.luan/audioplayers",
+        &StandardMethodCodec::GetInstance());
+    globalMethods = std::make_unique<MethodChannel<EncodableValue>>(
+        binaryMessenger, "xyz.luan/audioplayers.global",
+        &StandardMethodCodec::GetInstance());
+    auto _globalEventChannel = std::make_unique<EventChannel<EncodableValue>>(
+        binaryMessenger, "xyz.luan/audioplayers.global/events",
+        &StandardMethodCodec::GetInstance());
 
-  auto plugin = std::make_unique<AudioplayersWindowsPlugin>();
+    auto plugin = std::make_unique<AudioplayersWindowsPlugin>();
 
-  channel->SetMethodCallHandler(
-      [plugin_pointer = plugin.get()](const auto &call, auto result) {
-        plugin_pointer->HandleMethodCall(call, std::move(result));
-      });
+    methods->SetMethodCallHandler(
+        [plugin_pointer = plugin.get()](const auto &call, auto result) {
+            plugin_pointer->HandleMethodCall(call, std::move(result));
+        });
 
-  globalChannel->SetMethodCallHandler(
-      [plugin_pointer = plugin.get()](const auto &call, auto result) {
-        plugin_pointer->HandleGlobalMethodCall(call, std::move(result));
-      });
+    globalMethods->SetMethodCallHandler(
+        [plugin_pointer = plugin.get()](const auto &call, auto result) {
+            plugin_pointer->HandleGlobalMethodCall(call, std::move(result));
+        });
+    globalEvents = std::make_unique<EventStreamHandler<>>();
+    auto _obj_stm_handle =
+        static_cast<StreamHandler<EncodableValue> *>(globalEvents.get());
+    std::unique_ptr<StreamHandler<EncodableValue>> _ptr{_obj_stm_handle};
+    _globalEventChannel->SetStreamHandler(std::move(_ptr));
 
-  registrar->AddPlugin(std::move(plugin));
+    registrar->AddPlugin(std::move(plugin));
 }
 
 AudioplayersWindowsPlugin::AudioplayersWindowsPlugin() {}
@@ -97,134 +108,158 @@ AudioplayersWindowsPlugin::~AudioplayersWindowsPlugin() {}
 void AudioplayersWindowsPlugin::HandleGlobalMethodCall(
     const MethodCall<EncodableValue> &method_call,
     std::unique_ptr<MethodResult<EncodableValue>> result) {
+    auto args = method_call.arguments();
 
-  auto args = method_call.arguments();
-
-  if (method_call.method_name().compare("changeLogLevel") == 0) {
-    auto valueName = GetArgument<std::string>("value", args, std::string());
-    if(valueName.empty()) {
-        Logger::Error("Null value received on changeLogLevel");
-        result->Success(EncodableValue(0));
+    if (method_call.method_name().compare("setAudioContext") == 0) {
+        this->OnGlobalLog("Setting AudioContext is not supported on Windows");
+    } else if (method_call.method_name().compare("emitLog") == 0) {
+        auto message = GetArgument<std::string>("message", args, std::string());
+        this->OnGlobalLog(message);
+    } else if (method_call.method_name().compare("emitError") == 0) {
+        auto code = GetArgument<std::string>("code", args, std::string());
+        auto message = GetArgument<std::string>("message", args, std::string());
+        globalEvents->Error(code, message, nullptr);
+        result->Success(EncodableValue(1));
+    } else {
+        result->NotImplemented();
         return;
     }
-    LogLevel value;
-    if (valueName.compare("LogLevel.info") == 0) {
-      value = LogLevel::Info;
-    } else if (valueName.compare("LogLevel.error") == 0) {
-      value = LogLevel::Error;
-    } else if (valueName.compare("LogLevel.none") == 0) {
-      value = LogLevel::None;
-    } else {
-      Logger::Error("Invalid value received on changeLogLevel");
-      result->Success(EncodableValue(0));
-      return;
-    }
 
-    Logger::logLevel = value;
-  }
-
-  result->Success(EncodableValue(1));
+    result->Success(EncodableValue(1));
 }
 
 void AudioplayersWindowsPlugin::HandleMethodCall(
     const MethodCall<EncodableValue> &method_call,
     std::unique_ptr<MethodResult<EncodableValue>> result) {
+    auto args = method_call.arguments();
 
-  auto args = method_call.arguments();
-
-  auto playerId = GetArgument<std::string>("playerId", args, std::string());
-  if(playerId.empty()) {
-    Logger::Error("Call missing mandatory parameter playerId.");
-    result->Success(EncodableValue(0));
-  }
-  auto mode = GetArgument<std::string>("mode", args, std::string());
-  auto player = GetPlayer(playerId, mode);
-
-  if (method_call.method_name().compare("pause") == 0) {
-    player->Pause();
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("resume") == 0) {
-    player->Resume();
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("stop") == 0) {
-    player->Pause();
-    player->SeekTo(0);
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("release") == 0) {
-    player->Pause();
-    player->SeekTo(0);
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("seek") == 0) {
-    auto position = GetArgument<int>("position", args, (int)(player->GetPosition() / 10000));
-    player->SeekTo(static_cast<int64_t>(position * 10000.0));
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("setSourceUrl") == 0) {
-    auto url = GetArgument<std::string>("url", args, std::string());
-
-    if (url.empty()) {
-        Logger::Error("Null URL received on setSourceUrl");
-        result->Success(EncodableValue(0));
+    auto playerId = GetArgument<std::string>("playerId", args, std::string());
+    if (playerId.empty()) {
+        result->Error("WindowsAudioError", "Call missing mandatory parameter playerId.",
+                      nullptr);
         return;
     }
 
-    try {
-      player->SetSourceUrl(url);
-      result->Success(EncodableValue(1));
-    } catch(...) {
-      Logger::Error("Error setting url to '" + url + "'.");
-      result->Success(EncodableValue(0));
-    }
-  } else if (method_call.method_name().compare("getDuration") == 0) {
-    result->Success(EncodableValue(player->GetDuration() / 10000));
-  } else if (method_call.method_name().compare("setVolume") == 0) {
-    auto volume = GetArgument<double>("volume", args, 1.0);
-    player->SetVolume(volume);
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("getCurrentPosition") == 0) {
-    result->Success(EncodableValue(player->GetPosition() / 10000));
-  } else if (method_call.method_name().compare("setPlaybackRate") == 0) {
-    auto playbackRate = GetArgument<double>("playbackRate", args, 1.0);
-    player->SetPlaybackSpeed(playbackRate);
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("setReleaseMode") == 0) {
-    auto releaseMode = GetArgument<std::string>("releaseMode", args, std::string());
-    if (releaseMode.empty()) {
-        Logger::Error("Error calling setReleaseMode, releaseMode cannot be null");
-        result->Success(EncodableValue(0));
+    if (method_call.method_name().compare("create") == 0) {
+        CreatePlayer(playerId);
+        result->Success(EncodableValue(1));
         return;
     }
-    auto looping = releaseMode.find("loop") != std::string::npos;
-    player->SetLooping(looping);
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("setPlayerMode") == 0) {
-    // windows doesn't have multiple player modes, so this should no-op
-    result->Success(EncodableValue(1));
-  } else if (method_call.method_name().compare("setBalance") == 0) {
-      auto balance = GetArgument<double>("balance", args, 0.0);
-      player->SetBalance(balance);
-      result->Success(EncodableValue(1));
-  } else {
-    result->NotImplemented();
-  }
+
+    auto player = GetPlayer(playerId);
+
+    if (method_call.method_name().compare("pause") == 0) {
+        player->Pause();
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("resume") == 0) {
+        player->Resume();
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("stop") == 0) {
+        player->Pause();
+        player->SeekTo(0);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("release") == 0) {
+        player->Pause();
+        player->SeekTo(0);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("seek") == 0) {
+        auto position = GetArgument<int>("position", args,
+                                         (int)(player->GetPosition() / 10000));
+        player->SeekTo(static_cast<int64_t>(position * 10000.0));
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("setSourceUrl") == 0) {
+        auto url = GetArgument<std::string>("url", args, std::string());
+
+        if (url.empty()) {
+            result->Error("WindowsAudioError", "Null URL received on setSourceUrl", nullptr);
+            return;
+        }
+
+        try {
+            player->SetSourceUrl(url);
+            result->Success(EncodableValue(1));
+        } catch (...) {
+            result->Error("WindowsAudioError", "Error setting url to '" + url + "'.", nullptr);
+        }
+    } else if (method_call.method_name().compare("getDuration") == 0) {
+        result->Success(EncodableValue(player->GetDuration() / 10000));
+    } else if (method_call.method_name().compare("setVolume") == 0) {
+        auto volume = GetArgument<double>("volume", args, 1.0);
+        player->SetVolume(volume);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("getCurrentPosition") == 0) {
+        result->Success(EncodableValue(player->GetPosition() / 10000));
+    } else if (method_call.method_name().compare("setPlaybackRate") == 0) {
+        auto playbackRate = GetArgument<double>("playbackRate", args, 1.0);
+        player->SetPlaybackSpeed(playbackRate);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("setReleaseMode") == 0) {
+        auto releaseMode =
+            GetArgument<std::string>("releaseMode", args, std::string());
+        if (releaseMode.empty()) {
+            result->Error(
+                "WindowsAudioError", "Error calling setReleaseMode, releaseMode cannot be null",
+                nullptr);
+            return;
+        }
+        auto looping = releaseMode.find("loop") != std::string::npos;
+        player->SetLooping(looping);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("setPlayerMode") == 0) {
+        // windows doesn't have multiple player modes, so this should no-op
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("setBalance") == 0) {
+        auto balance = GetArgument<double>("balance", args, 0.0);
+        player->SetBalance(balance);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("emitLog") == 0) {
+        auto message = GetArgument<std::string>("message", args, std::string());
+        player->OnLog(message);
+        result->Success(EncodableValue(1));
+    } else if (method_call.method_name().compare("emitError") == 0) {
+        auto code = GetArgument<std::string>("code", args, std::string());
+        auto message = GetArgument<std::string>("message", args, std::string());
+        player->OnError(code, message, nullptr);
+        result->Success(EncodableValue(1));
+    } else {
+        result->NotImplemented();
+    }
 }
 
-AudioPlayer* AudioplayersWindowsPlugin::GetPlayer(std::string playerId, std::string mode) {
-  auto searchPlayer = audioPlayers.find(playerId);
-  if(searchPlayer != audioPlayers.end()) {
-    return searchPlayer->second.get();
-  } else {
-    auto player = std::make_unique<AudioPlayer>(playerId, channel.get());
-    auto playerPtr = player.get();
+void AudioplayersWindowsPlugin::CreatePlayer(std::string playerId) {
+    auto eventChannel = std::make_unique<EventChannel<EncodableValue>>(
+        binaryMessenger, "xyz.luan/audioplayers/events/" + playerId,
+        &StandardMethodCodec::GetInstance());
+
+    auto eventHandler = new EventStreamHandler<>();
+    auto _obj_stm_handle =
+        static_cast<StreamHandler<EncodableValue> *>(eventHandler);
+    std::unique_ptr<StreamHandler<EncodableValue>> _ptr{_obj_stm_handle};
+    eventChannel->SetStreamHandler(std::move(_ptr));
+
+    auto player =
+        std::make_unique<AudioPlayer>(playerId, methods.get(), eventHandler);
     audioPlayers.insert(std::make_pair(playerId, std::move(player)));
-    return playerPtr;
-  }
+}
+
+AudioPlayer *AudioplayersWindowsPlugin::GetPlayer(std::string playerId) {
+    auto searchPlayer = audioPlayers.find(playerId);
+    return searchPlayer->second.get();
+}
+
+void AudioplayersWindowsPlugin::OnGlobalLog(const std::string& message) {
+    globalEvents->Success(std::make_unique<flutter::EncodableValue>(
+        flutter::EncodableMap({{flutter::EncodableValue("event"),
+                                flutter::EncodableValue("audio.onLog")},
+                               {flutter::EncodableValue("value"),
+                                flutter::EncodableValue(message)}})));
 }
 
 }  // namespace
 
 void AudioplayersWindowsPluginRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar) {
-  AudioplayersWindowsPlugin::RegisterWithRegistrar(
-      PluginRegistrarManager::GetInstance()
-          ->GetRegistrar<PluginRegistrarWindows>(registrar));
+    AudioplayersWindowsPlugin::RegisterWithRegistrar(
+        PluginRegistrarManager::GetInstance()
+            ->GetRegistrar<PluginRegistrarWindows>(registrar));
 }
